scorer/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
pyarrow==14.0.2
catboost==1.2.5
joblib==1.3.2
PyYAML==6.0.2
aiokafka==0.10.0